[2단계 - 장바구니 협업] 자스민(황정민) 미션 제출합니다. (#53)
* feat: withAuthPage 대신에 PrivateWrapper, PublicWrapper 이용하여 페이지 인증 처리구현

* rename: AuthPage -> AuthForm 으로 이름 변경

* feat: api 명세에 따른 코드 수정

* fix: AUTH_BASE_URL -> BASE_URL로 통합

* feat: api 명세에 맞도록 thunk 변경

* feat: msw 코드 삭제

* feat: 사용하지않는 인자 삭제

* fix: 로컬 실행을 위해 .env 주소를 변경

* feat: api 주소들 주석으로 명세

+++ src/redux/cartList/thunk.ts
@@ -1,14 +1,19 @@
-import { BASE_URL } from 'apis';
-import axios from 'axios';
 import type { Dispatch } from 'redux';
 import { CartListAction, cartListActions } from 'redux/cartList/action';
-import { RootState } from 'redux/rootReducer';
-import { CartItem } from 'types/domain';
+import { client } from 'apis';
 
 export const getCartListRequest = () => async (dispatch: Dispatch<CartListAction>) => {
+  const accessToken = localStorage.getItem('access-token');
+
+  if (!accessToken) return;
+
   dispatch(cartListActions.getCartListActionGroup.request());
   try {
-    const response = await axios.get(`${BASE_URL}/cartList`);
+    const response = await client.get('/customers/carts', {
+      headers: {
+        Authorization: `Bearer ${accessToken}`,
+      },
+    });
 
     dispatch(cartListActions.getCartListActionGroup.success(response.data));
   } catch (e: unknown) {
@@ -19,10 +24,23 @@ export const getCartListRequest = () => async (dispatch: Dispatch<CartListAction
 };
 
 export const putCartItemRequest =
-  (cartItem: CartItem) => async (dispatch: Dispatch<CartListAction>) => {
+  ({ id, quantity }: { id: number; quantity: number }) =>
+  async (dispatch: Dispatch<CartListAction>) => {
+    const accessToken = localStorage.getItem('access-token');
+
+    if (!accessToken) return;
+
     dispatch(cartListActions.putCartItemActionGroup.request());
     try {
-      const response = await axios.put(`${BASE_URL}/cartList/${cartItem.id}`, cartItem);
+      const response = await client.put(
+        `customers/carts/${id}`,
+        { quantity },
+        {
+          headers: {
+            Authorization: `Bearer ${accessToken}`,
+          },
+        }
+      );
 
       dispatch(cartListActions.putCartItemActionGroup.success(response.data));
     } catch (e: unknown) {
@@ -33,10 +51,22 @@ export const putCartItemRequest =
   };
 
 export const postCartItemRequest =
-  (cartItem: CartItem) => async (dispatch: Dispatch<CartListAction>) => {
+  (productId: number) => async (dispatch: Dispatch<CartListAction>) => {
+    const accessToken = localStorage.getItem('access-token');
+
+    if (!accessToken) return;
+
     dispatch(cartListActions.postCartItemActionGroup.request());
     try {
-      const response = await axios.post(`${BASE_URL}/cartList`, cartItem);
+      const response = await client.post(
+        '/customers/carts',
+        { productId },
+        {
+          headers: {
+            Authorization: `Bearer ${accessToken}`,
+          },
+        }
+      );
 
       dispatch(cartListActions.postCartItemActionGroup.success(response.data));
     } catch (e: unknown) {
@@ -46,46 +76,18 @@ export const postCartItemRequest =
     }
   };
 
-export const patchCartSelectedRequest =
-  (id: number) => async (dispatch: Dispatch<CartListAction>, getState: () => RootState) => {
-    dispatch(cartListActions.patchCartSelectedActionGroup.request());
-    try {
-      const { data: cartList } = getState().cartList;
-      const isSelected = cartList.find(item => item.id === id).isSelected;
-      const response = await axios.patch(`${BASE_URL}/cartList/${id}`, { isSelected: !isSelected });
-
-      dispatch(cartListActions.patchCartSelectedActionGroup.success(response.data));
-    } catch (e: unknown) {
-      if (e instanceof Error) {
-        dispatch(cartListActions.patchCartSelectedActionGroup.failure(e));
-      }
-    }
-  };
-
-export const patchAllCartSelectedRequest =
-  (isAllSelected: boolean) =>
-  async (dispatch: Dispatch<CartListAction>, getState: () => RootState) => {
-    dispatch(cartListActions.patchAllCartSelectedActionGroup.request());
-    try {
-      const { data: cartList } = getState().cartList;
-
-      cartList.forEach(
-        async item =>
-          await axios.patch(`${BASE_URL}/cartList/${item.id}`, { isSelected: !isAllSelected })
-      );
-
-      dispatch(cartListActions.patchAllCartSelectedActionGroup.success(!isAllSelected));
-    } catch (e: unknown) {
-      if (e instanceof Error) {
-        dispatch(cartListActions.patchAllCartSelectedActionGroup.failure(e));
-      }
-    }
-  };
-
 export const deleteCartItemRequest = (id: number) => async (dispatch: Dispatch<CartListAction>) => {
+  const accessToken = localStorage.getItem('access-token');
+
+  if (!accessToken) return;
+
   dispatch(cartListActions.deleteCartItemActionGroup.request());
   try {
-    await axios.delete(`${BASE_URL}/cartList/${id}`);
+    await client.delete(`/customers/carts/${id}`, {
+      headers: {
+        Authorization: `Bearer ${accessToken}`,
+      },
+    });
 
     dispatch(cartListActions.deleteCartItemActionGroup.success(id));
   } catch (e: unknown) {
@@ -95,18 +97,23 @@ export const deleteCartItemRequest = (id: number) => async (dispatch: Dispatch<C
   }
 };
 
-export const deleteAllCartItemRequest =
-  () => async (dispatch: Dispatch<CartListAction>, getState: () => RootState) => {
-    dispatch(cartListActions.deleteAllCartItemActionGroup.request());
-    try {
-      const { data: cartList } = getState().cartList;
+export const deleteAllCartItemRequest = () => async (dispatch: Dispatch<CartListAction>) => {
+  const accessToken = localStorage.getItem('access-token');
 
-      cartList.forEach(async item => await axios.delete(`${BASE_URL}/cartList/${item.id}`));
+  if (!accessToken) return;
 
-      dispatch(cartListActions.deleteAllCartItemActionGroup.success());
-    } catch (e: unknown) {
-      if (e instanceof Error) {
-        dispatch(cartListActions.deleteAllCartItemActionGroup.failure(e));
-      }
+  dispatch(cartListActions.deleteAllCartItemActionGroup.request());
+  try {
+    await client.delete('/customers/carts', {
+      headers: {
+        Authorization: `Bearer ${accessToken}`,
+      },
+    });
+
+    dispatch(cartListActions.deleteAllCartItemActionGroup.success());
+  } catch (e: unknown) {
+    if (e instanceof Error) {
+      dispatch(cartListActions.deleteAllCartItemActionGroup.failure(e));
     }
-  };
+  }
+};
